requirement.txt
diff-match-patch==20121119
Django==2.0.5
django-import-export==1.0.1
et-xmlfile==1.0.1
future==0.16.0
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
vertica-python==0.7.3
xlrd==1.1.0
xlwt==1.3.0
